requirements.txt
Cython
numpy
scipy
matplotlib
mpmath
pandas
seaborn
numba
pyyaml
isort
flake8
flake8-isort